geodjango form support
