+++ grumpy/books/templates/books/book_form.html
@@ -24,12 +24,18 @@
             <form action="." method="post">
                 {% csrf_token %}
 
-                {{ form.non_field_errors }}
+                <div class="text-danger row">
+                    {{ form.non_field_errors }}
+                </div>
 
                 {% for field in form %}
-                    <div class="form-group">
-                        {{ field.errors }}
-                        {{ field.label_tag }} {{ field }}
+                    <div class="form-group row">
+                        <div class="col-sm-2">{{ field.label_tag }}</div>
+                        <div class="col-sm-10">
+                            {{ field }}
+                            <div class="help form-text text-muted">{{ field.field.help_text|safe }}</div>
+                            <div class="text-danger">{{ field.errors }}</div>
+                        </div>
                     </div>
                 {% endfor %}
