+++ grumpy/core/static/core/css/base.css
@@ -1,16 +1,28 @@
 #grumpy div#banner {
-  background-color: #150904;
+    background-color: #150904;
 
 }
 #grumpy .footer {
-  /* position: absolute; */
-  margin-top: 2px;
-  bottom: 0;
-  width: 100%;
-  background-color: #f5f5f5;
+    /* position: absolute; */
+    margin-top: 2px;
+    bottom: 0;
+    width: 100%;
+    background-color: #f5f5f5;
 }
 
 #grumpy a.unstyled {
-  color: inherit;
-  text-decoration: none !important;
+    color: inherit;
+    text-decoration: none !important;
+}
+
+#grumpy ul.errorlist {
+    list-style-type: none;
+    padding-left: 0;
+}
+
+.dj_map {
+    /* known bug in geodjango */
+    /* they don't properly support OSMWidget */
+    width: 400px;
+    height: 400px;
 }
